HUB ETL BASE PDFs/Parser_PDF_Hospitais_Municipais/src/requirements.txt
pandas
tabula-py
openpyxl
numpy